Re-structured the code completely to use modules.
This will be better in the long run for code portability as we now use a base.rb file to track overall structure and leave a large portion of the logic to the individual files. Also introduced the `::QuoraWebDriver` namespace to clean up more code. We don't want to clutter that shit.